Removed ship direction, it solves many potential problems: 1 - What if we were hitting ship 1, and we hit ship 2, i.e:
ship1 (not hit)--- emptySpace
ship 1(hit)---------ship2(hit) ----ship2(not hit)

The algorithm will destroy ship2, but due to different
algorithms like shortesShip factor for finding new ships -
it may not be efficient at finding ship1 as it is already hit
and it's space is already covered. So if the shortest ship factor
was 5, but ship 1 was hit, the board could even run
out of space before begining to hit ship 1.

2 - (related to pt 1), what happens if we hit another ship?
Actions allow us to trail both.

3 - Implementation issues with shipDirection: when do we stop?
    Actions give us two conditions:
        i - we ran out of Actions, meaning we could even pursuit the second
        ship potentially
        ii - we sunk a ship